requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
scikit-learn==1.5.2
numpy==2.1.2
mysql-connector-python==9.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
gunicorn==23.0.0